Moved files to utils module
