+++ src/openfermion/measurements/__init__.py
@@ -17,14 +17,3 @@ from ._equality_constraint_projection import (apply_constraints,
 
 from ._rdm_equality_constraints import (one_body_fermion_constraints,
                                         two_body_fermion_constraints)
-
-from ._rdm_mapping_functions import (kronecker_delta,
-                                     map_two_pdm_to_two_hole_dm,
-                                     map_two_pdm_to_one_pdm,
-                                     map_one_pdm_to_one_hole_dm,
-                                     map_one_hole_dm_to_one_pdm,
-                                     map_two_pdm_to_particle_hole_dm,
-                                     map_two_hole_dm_to_two_pdm,
-                                     map_two_hole_dm_to_one_hole_dm,
-                                     map_particle_hole_dm_to_one_pdm,
-                                     map_particle_hole_dm_to_two_pdm)
